adicionando criptografia nas senhas

+++ app.py
@@ -2,6 +2,7 @@ from flask import Flask, request, jsonify
 from models.user import User
 from database import db
 from flask_login import LoginManager, login_user, logout_user, login_required, current_user
+import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "YOUR_SECRET_KEY"
@@ -12,11 +13,13 @@ db.init_app(app)
 login_manager.init_app(app)
 
 login_manager.login_view = 'login'
-
 @login_manager.user_loader
 def load_user(user_id):
     return User.query.get(user_id)
 
+###########################
+## LOGIN
+###########################
 @app.route('/login', methods=['POST'])
 def login():
     data = request.json
@@ -26,19 +29,25 @@ def login():
     if username and password:
         user = User.query.filter_by(username=username).first()
 
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             print(current_user.is_authenticated)
             return jsonify({"message": "Autenticação realizada com sucesso"}), 200
 
     return jsonify({"message": "Credênciais inválidas"}), 400
 
+###########################
+## LOGOUT
+###########################
 @app.route('/logout', methods=['GET'])
 @login_required
 def logout():
     logout_user()
     return jsonify({"message": "Logout realizado com sucesso"}), 200
 
+###########################
+## CREATE
+###########################
 @app.route('/user', methods=['POST'])
 def create_user():
     data = request.json
@@ -46,13 +55,17 @@ def create_user():
     password = data.get("password")
     
     if username and password:
-        user = User(username=username, password=password)
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, password=hashed_password, role='user')
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuário cadastrado com sucesso"}), 200
     
     return jsonify({"message": "Dados inválidos"}), 400
 
+###########################
+## READ
+###########################
 @app.route('/user/<int:id_user>', methods=['GET'])
 @login_required
 def read_user(id_user):
@@ -63,11 +76,17 @@ def read_user(id_user):
     
     return jsonify({"message": "Usuário não encontrado"}), 404
 
+###########################
+## UPDATE
+###########################
 @app.route('/user/<int:id_user>', methods=['PUT'])
 @login_required
 def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
+    
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "Operação não permitida"}), 404
     
     if user and data.get("password"):
         user.password = data.get("password")
@@ -77,16 +96,20 @@ def update_user(id_user):
 
     return jsonify({"message": "Usuário não encontrado"}), 404
 
-
+###########################
+## DELETE
+###########################
 @app.route('/user/<int:id_user>', methods=['DELETE'])
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
     
+    if current_user.role != "admin":
+        return jsonify({"message": "Operação não permitida"}), 403
+    
     if id_user == current_user.id:
         return jsonify({"message": "Deleção não permitida"}), 403
 
-    
     if user:
         db.session.delete(user)
         db.session.commit()
